pubsub-publisher/requirements.txt
annotated-types==0.6.0 ; python_version >= "3.8" and python_version < "4.0"
anyio==3.7.1 ; python_version >= "3.8" and python_version < "4.0"
cachetools==5.3.2 ; python_version >= "3.8" and python_version < "4.0"
certifi==2023.7.22 ; python_version >= "3.8" and python_version < "4.0"
charset-normalizer==3.3.2 ; python_version >= "3.8" and python_version < "4.0"
click==8.1.7 ; python_version >= "3.8" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.8" and python_version < "4.0" and platform_system == "Windows"
exceptiongroup==1.1.3 ; python_version >= "3.8" and python_version < "3.11"
fastapi==0.104.1 ; python_version >= "3.8" and python_version < "4.0"
google-api-core==2.13.0 ; python_version >= "3.8" and python_version < "4.0"
google-api-core[grpc]==2.13.0 ; python_version >= "3.8" and python_version < "4.0"
google-auth==2.23.4 ; python_version >= "3.8" and python_version < "4.0"
google-cloud-bigquery==3.13.0 ; python_version >= "3.8" and python_version < "4.0"
google-cloud-core==2.3.3 ; python_version >= "3.8" and python_version < "4.0"
google-cloud-pubsub==2.18.4 ; python_version >= "3.8" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.8" and python_version < "4.0"
google-resumable-media==2.6.0 ; python_version >= "3.8" and python_version < "4.0"
googleapis-common-protos==1.61.0 ; python_version >= "3.8" and python_version < "4.0"
googleapis-common-protos[grpc]==1.61.0 ; python_version >= "3.8" and python_version < "4.0"
grpc-google-iam-v1==0.12.6 ; python_version >= "3.8" and python_version < "4.0"
grpcio-status==1.59.2 ; python_version >= "3.8" and python_version < "4.0"
grpcio==1.59.2 ; python_version >= "3.8" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.8" and python_version < "4.0"
idna==3.4 ; python_version >= "3.8" and python_version < "4.0"
packaging==23.2 ; python_version >= "3.8" and python_version < "4.0"
proto-plus==1.22.3 ; python_version >= "3.8" and python_version < "4.0"
protobuf==4.25.0 ; python_version >= "3.8" and python_version < "4.0"
pyasn1-modules==0.3.0 ; python_version >= "3.8" and python_version < "4.0"
pyasn1==0.5.0 ; python_version >= "3.8" and python_version < "4.0"
pydantic-core==2.10.1 ; python_version >= "3.8" and python_version < "4.0"
pydantic==2.4.2 ; python_version >= "3.8" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.8" and python_version < "4.0"
requests==2.31.0 ; python_version >= "3.8" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.8" and python_version < "4"
six==1.16.0 ; python_version >= "3.8" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.8" and python_version < "4.0"
starlette==0.27.0 ; python_version >= "3.8" and python_version < "4.0"
typing-extensions==4.8.0 ; python_version >= "3.8" and python_version < "4.0"
urllib3==2.0.7 ; python_version >= "3.8" and python_version < "4.0"
uvicorn==0.24.0.post1 ; python_version >= "3.8" and python_version < "4.0"
